requirements.txt
Flask==2.1.1
boto3==1.24.10
python-dotenv==0.20.0
nodemon